rrr: convert rpc/frame_codec to module (rrr.frame_codec)
frame_codec.cpp becomes the rrr.frame_codec module; .hpp shim.
Exports CodecError enum + framing helpers (encode_request_frame /
encode_response_frame / decode_request_frame / decode_response_frame
plus the underlying byte-layout utilities). Imports
rrr.internal_protocol for the response-header extension flag
constants.

36 modules. 81.47 s wallclock = 4.83×.
